requirements.txt
requests
spotipy
streamlit
numpy
pandas
matplotlib
plotly
pillow
fastai
torch
typing
librosa
pydub